+++ proknow/Patients/__init__.py
@@ -1,3 +1,4 @@
 from .Entities import *
 from .Studies import *
+from .Scorecards import *
 from .Patients import *